// GMAOWB/src/app/Models/Piece.ts
export class Piece {
    CodePiece: string;
    Designation: string;
    Unite: string;
    NumSerie: string;
    PAchat: number;
    CodeFournisseur: string;
    StockMin: number;
    StockMax: number;
    StockAlerte: number;
    QteStock :number;
    AnCode: string;
    Machine: string;
    CodeCC:string;
    Previent: number;
    RefFournisseur:string;
    Analyse1: string;
    Analyse2: string;
    Analyse3: string;
    PAchatTTC: number;
    ZStockage: string;
    IF_CC: string;
    CentreCout:string;
    Marque:string;
    AnneeConstruction:number;
    DateMes: Date;
    Etalonnage:string;
    Image:string;
    FicheT: string;
    QteCmd:number;
    TestCCout:string;
    DateAchat:Date;
    DateFabrication:Date;
    Obsolete:string;

  }